+++ ksef-client/src/main/java/pl/akmf/ksef/sdk/client/model/ApiException.java
@@ -3,14 +3,18 @@ package pl.akmf.ksef.sdk.client.model;
 import java.net.http.HttpHeaders;
 import java.util.stream.Collectors;
 
-public class ApiException extends Exception {
+public abstract class ApiException extends Exception {
     private final int code;
+    private final String url;
+    private final String method;
     private final transient HttpHeaders responseHeaders;
     private final transient ExceptionResponse exceptionResponse;
 
     public ApiException(int code, String message) {
         super(message);
         this.code = code;
+        this.url = "";
+        this.method = "";
         this.responseHeaders = null;
         this.exceptionResponse = null;
     }
@@ -18,6 +22,8 @@ public class ApiException extends Exception {
     public ApiException(Throwable throwable) {
         super(throwable);
         this.code = 0;
+        this.url = "";
+        this.method = "";
         this.responseHeaders = null;
         this.exceptionResponse = null;
     }
@@ -25,6 +31,8 @@ public class ApiException extends Exception {
     public ApiException(String message) {
         super(message);
         this.code = 0;
+        this.url = "";
+        this.method = "";
         this.responseHeaders = null;
         this.exceptionResponse = null;
     }
@@ -32,12 +40,31 @@ public class ApiException extends Exception {
     public ApiException(int code, String message, HttpHeaders responseHeaders, ExceptionResponse exceptionResponse) {
         super(message);
         this.code = code;
+        this.url = "";
+        this.method = "";
+        this.exceptionResponse = exceptionResponse;
+        this.responseHeaders = responseHeaders;
+    }
+
+    public ApiException(int code, String url, String method, String message, HttpHeaders responseHeaders, ExceptionResponse exceptionResponse) {
+        super(message);
+        this.code = code;
+        this.url = url;
+        this.method = method;
         this.exceptionResponse = exceptionResponse;
         this.responseHeaders = responseHeaders;
     }
 
     public int getCode() {
         return code;
+    }
+
+    public String getUrl() {
+        return url;
+    }
+
+    public String getMethod() {
+        return method;
     }
 
     public ExceptionResponse getExceptionResponse() {
@@ -50,15 +77,15 @@ public class ApiException extends Exception {
 
     @Override
     public String toString() {
-        return "ApiException{" +
-                "\ncode=" + code +
+        return "code=" + code +
+                "\nurl=" + url +
+                "\nmethod=" + method +
                 ",\nresponseHeaders=" + (responseHeaders != null
                 ? responseHeaders.map().entrySet().stream()
                 .flatMap(entry -> entry.getValue().stream()
                         .map(value -> "'" + entry.getKey() + ": " + value + "'"))
                 .collect(Collectors.joining(", ")) : responseHeaders) +
                 ",\n" + exceptionResponse +
-                ",\nmessage=" + getMessage() +
-                "\n}";
+                ",\nmessage=" + getMessage();
     }
 }